Add stage name customization
In some deployments, per the [12-Factor App manifesto][], the same
`Rails.env` is used for staging, QA, demo or any other production-like
deployment.

Enable the name of the stage that is added to the prefix to be
configured to accomodate these scenarios where the deploy environment is
`staging` but `Rails.env` is `production`.

[12-Factor App manifesto]: http://12factor.net/config